Add column selection button
Related Issues: #23

A new button has been added above the molecule table which allows users
to select the value collections which are displayed in the table. If
the value collection is toggled off, it is also not possible for the
user to sort by this collection. The idea is that this will make the
drop down list smaller for the sorting list, so that it is easier for
users to find the collection they want to sort by.

The implementation through the introduction of two new actions,
`HideCollection` and `ShowCollection`, which add and remove a given
collection from the `ignoredCollections` list. The `ignoredCollections`
list is then used to filter which collections in `valueCollections` are
displayed in the table, as well as the sorting drop-down menu.
The `ignoredCollections` state is passed around between the different
pages. This is because I expect that users would be quite annoyed
at having to continuously re-select which columns they wish to view
upon changing the page.

The `ignoredCollections` are also no longer being sent to requests.
The `ignoredCollections` sent to requests are now pretty much empty.
Conceptually the `ignoredCollections` used by the page states and by
the requests are quite different. The page's `ignoredCollections`
are those which are not currently being viewed by the table. However,
for these we still want to perform data requests. The reason being
that if the user decides to turn the column back on, there is no need
to send a new request to see the values. In addition, if these
collections were passed to the request, when the request returns the
new value collections, the `ignoredCollections` would not be present.
This would mean that you would not be able to show the empty checkbox,
because that depends on the collection being present in value
collections. This could be avoided by going through all the values
in `ignoredCollections` too, and making an empty checkbox for those,
but I think that leads to a more complex implementation. Essentially,
I think the `ignoredCollections` in the page state vs the requests
serve different purposes, and it's a bad idea to couple them.

There is a performance trade-off to not coupling the request and
page `ignoredCollections`. If they were coupled, then the page
requests would likely be quicker, because there would be fewer data
queries. However, toggling of the columns would be more expensive, as
it would involve requests. Because I think users expect toggling of
columns to be snappy, while page loads to be slower, I think not
coupling them will annoy them the least.

Alos, by default, the various position matrix collections were being
added into `ignoredCollections`, this is no longer the case. I'm
just going to expect the users toggle the collection off explicitly
if they don't want to see it. I think doing it behind the scenes is
a little bit confusing. I could see adding it back, but I don't really
want to think about it right now. Really I think the collection would
need to be added into the `ignoredCollections` for the requests,
rather than for the page state. This would prevent the collection
showing up as a `valueCollection`, once the request returns the new
list of `valueCollections`. I guess I'm not too bothered by this,
because I expect there will be multiple position matrix collections
which the user will need to toggle off. For example, if the user
has 5 position matrix collections for building blocks. They can
put at most 1 as a building block position matrix collection into
the Mongo Configurator. So I could automatically hide this one, but
the user would have to manually hide the others. So having the extra
one there is not a big deal and maybe even slightly more consistent.

I also changed lots of types from `Array String` to `HashSet String`
for better performance and to be more explicit that there are
unsorted collections of unique values